requirements.txt
Django==2.1
django-debug-toolbar==1.8
gunicorn==19.4.5
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==3.3.1
psycopg2
